refactor: removed duplicated fields and addresses
