+++ yarn-project/circuits.js/package.json
@@ -35,6 +35,7 @@
     "rootDir": "./src"
   },
   "dependencies": {
+    "@aztec/foundation": "workspace:^",
     "@aztec/wasm": "workspace:^",
     "cross-fetch": "^3.1.5",
     "detect-node": "^2.1.0",
@@ -42,7 +43,6 @@
     "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@aztec/foundation": "workspace:^",
     "@jest/globals": "^29.4.3",
     "@types/detect-node": "^2.0.0",
     "@types/jest": "^29.4.0",
