+++ yarn-project/circuits.js/src/structs/private_call_stack_item.ts
@@ -1,7 +1,7 @@
+import { AztecAddress } from '@aztec/foundation';
 import { serializeToBuffer } from '../utils/serialize.js';
 import { FunctionData } from './function_data.js';
 import { PrivateCircuitPublicInputs } from './private_circuit_public_inputs.js';
-import { AztecAddress } from './shared.js';
 
 /**
  * Call stack item on a private call.
