+++ yarn-project/foundation/src/index.ts
@@ -1,6 +1,9 @@
 export * from './async-map/index.js';
+export * from './aztec-address/index.js';
 export * from './bigint-buffer/index.js';
 export * from './errors/index.js';
+export * from './eth-address/index.js';
+export * from './fields/index.js';
 export * from './fifo/index.js';
 export * from './mutex/index.js';
 export * from './timer/index.js';
