+++ yarn-project/foundation/src/serialize/index.ts
@@ -1,3 +1,4 @@
 export * from './free_funcs.js';
 export * from './deserializer.js';
 export * from './serializer.js';
+export * from './buffer_reader.js';
